requirements.txt
discord.py==1.2.5
PyGithub==1.44.1
python-dotenv==0.10.3
